// src/app/models/Document Request/documentRequestLine.model.ts
export class DocumentRequestLine {

    TableNumber: any;
    HRRequestCode: any;

    DocumentRequestAddressCode: any;
    DocumentRequestTypeCode: any;
    
    DocumentLanguage: any;
    IncludeSalary: any;
    ConsulateType: any;
    CertificateType: any;
    Purpose: any;
    Remarks: any;
    LineNumber: any;
    IsDeleted: any;
    DataArea:any;
    ERPRecID:any;

    constructor() {

    }
}